+++ src/components/main/catalog/catalog.tsx
@@ -1,7 +1,8 @@
 import FilmCardsList from '../../film-list/film-cards-list';
 import { useAppDispatch, useAppSelector } from '../../../hooks';
-import classNames from 'classnames';
-import { changeFilterGenreAction, filterFilms } from '../../../actions/action';
+import { filterFilms } from '../../../actions/action';
+import { GenreList } from './genre-list';
+import { useEffect } from 'react';
 
 type Props = {
   genres: Set<string>;
@@ -10,32 +11,18 @@ type Props = {
 export function Catalog({ genres }: Props) {
   const dispatch = useAppDispatch();
 
-  const selectedGenre = useAppSelector((store) => store.genre);
   const films = useAppSelector((store) => store.films);
+  const selectedGenre = useAppSelector((store) => store.genre);
+
+  useEffect(() => {
+    dispatch(filterFilms());
+  }, [dispatch, selectedGenre]);
 
   return (
     <section className="catalog">
       <h2 className="catalog__title visually-hidden">Catalog</h2>
 
-      <ul className="catalog__genres-list">
-        {[...genres].map((genre) => {
-          const liClass = classNames({
-            'catalog__genres-item': true,
-            'catalog__genres-item--active': selectedGenre === genre
-          });
-
-          return (
-            <li className={liClass} key={genre}>
-              <a className='catalog__genres-link' onClick={() => {
-                dispatch(changeFilterGenreAction(genre));
-                dispatch(filterFilms());
-              }}
-              >{genre}
-              </a>
-            </li>
-          );
-        })}
-      </ul>
+      <GenreList genres={genres} />
 
       <FilmCardsList films={films} />
